+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [Unreleased]
+* CXV1-9744 Logging module error response unit test
 * CXV1-9841 Every 30 seconds, push logs to logging pipeline
 * CXV1-10071 Silent monitoring signal support w/ unit tests
 * Added additional logging around incoming MQTT messaging messages